// MochiR.Client/src/app/api/models/review-summary-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ReviewRatingDto } from '../models/review-rating-dto';
import { ReviewStatus } from '../models/review-status';
export interface ReviewSummaryDto {
  authorAvatarUrl: string | null;
  authorDisplayName: string | null;
  authorUserName: string | null;
  content: string | null;
  createdAt: string;
  excerpt: string | null;
  excerptIsAuto: boolean;
  id: number;
  isLikedByCurrentUser: boolean;
  likeCount: number;
  ratings: Array<ReviewRatingDto>;
  status: ReviewStatus;
  subjectId: number;
  subjectName: string | null;
  tags: Array<string>;
  title: string | null;
  userId: string;
}