// MochiR.Client/src/app/api/models/feed-item-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface FeedItemDto {
  authorAvatarUrl: string | null;
  authorDisplayName: string | null;
  authorId: string | null;
  authorUserName: string | null;
  content: string | null;
  createdAtUtc: string;
  reviewId: number;
  subjectId: number;
  subjectName: string | null;
  subjectSlug: string | null;
  subjectTypeId: number | null;
  title: string | null;
}